Update controller test for keys in response
This is a somewhat more aggressive test than before, on two counts. The
original test is extended to ensure that every pet object in the
response array has exactly the set of keys we want.

There is another test added as well, which asserts that every key in
each object has a non-nil value. This is certainly not appropriate for
every scenario, so if we want to not include it in this example project
I'll pull it out.